requirements.txt
numpy==1.17.5
pillow==6.2.2
flask==1.1.2
werkzeug==1.0.1
torch===1.5.0
torchvision===0.6.0
setuptools==46.1.3